chg: Refactor utilities into a separate utils package
