+++ flask_rebar/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals
 
-from flask_rebar.request_utils import marshal, response
+from flask_rebar.utils.request_utils import marshal, response
 
 from flask_rebar.rebar import (
     Rebar,
